pyproject.toml
[tool.poetry]
name = "oz06"
version = "0.1.0"
description = ""
authors = ["oz-01-006"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
pymongo = "^4.6.3"
pillow = "^10.3.0"
django-allauth = "^0.61.1"
mongoengine = "^0.28.2"
djoser = "^2.2.2"
djangorestframework = "^3.15.1"
django-rest-framework-mongoengine = "^3.4.1"
mysqlclient = "^2.2.4"
django-dotenv = "^1.4.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
environ = "^1.0"
django-cors-headers = "^4.3.1"
django-colorfield = "^0.11.0"
whitenoise = "^6.6.0"
common = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
